refactor: conform comments of pull request
