+++ packages/core/src/Background.ts
@@ -1,15 +1,6 @@
 import { Color } from "@oasis-engine/math";
+import { BackgroundMode } from "./enums/BackgroundMode";
 import { Sky } from "./sky/Sky";
-
-/**
- * The Background mode enumeration.
- */
-export enum BackgroundMode {
-  /* Solid color. */
-  Color,
-  /* Sky. */
-  Sky
-}
 
 /**
  * Background of scene.
@@ -20,11 +11,13 @@ export class Background {
    * @defaultValue `BackgroundMode.Sky`
    */
   mode: BackgroundMode = BackgroundMode.Sky;
+
   /**
    * Background color.
    * @defaultValue `new Color(0.25, 0.25, 0.25, 1.0)`
    */
   color: Color = new Color(0.25, 0.25, 0.25, 1.0);
+
   /*
    * Background sky.
    */
